refactor: move non readonly methods to trait
