+++ app/Policies/ReadOnlyPolicy.php
@@ -2,7 +2,6 @@
 
 namespace App\Policies;
 
-use App\Models\User\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
 /**
@@ -11,6 +10,7 @@ use Illuminate\Auth\Access\HandlesAuthorization;
 class ReadOnlyPolicy
 {
     use HandlesAuthorization;
+    use RejectsNonReadOnlyActions;
 
     public function view(): bool
     {
@@ -20,75 +20,5 @@ class ReadOnlyPolicy
     public function viewAny(): bool
     {
         return true;
-    }
-
-    public function attach(): bool
-    {
-        return false;
-    }
-
-    public function detach(): bool
-    {
-        return false;
-    }
-
-    public function update(): bool
-    {
-        return false;
-    }
-
-    public function create(): bool
-    {
-        return false;
-    }
-
-    public function delete(): bool
-    {
-        return false;
-    }
-
-    public function restore(): bool
-    {
-        return false;
-    }
-
-    public function forceDelete(): bool
-    {
-        return false;
-    }
-
-    public function detachAny(): bool
-    {
-        return false;
-    }
-
-    public function dissociate(): bool
-    {
-        return false;
-    }
-
-    public function dissociateAny(): bool
-    {
-        return false;
-    }
-
-    public function replicate(): bool
-    {
-        return false;
-    }
-
-    public function restoreAny(): bool
-    {
-        return false;
-    }
-
-    public function deleteAny(): bool
-    {
-        return false;
-    }
-
-    public function forceDeleteAny(): bool
-    {
-        return false;
     }
 }
