+++ app/Policies/ReadOnlyWithRolePolicy.php
@@ -8,10 +8,11 @@ use App\Models\User\User;
 /**
  * Policy that allows read only access only to users with a role.
  */
-class ReadOnlyWithRolePolicy extends ReadOnlyPolicy
+class ReadOnlyWithRolePolicy
 {
     use HandlesAuthorization;
     use ChecksUserRoles;
+    use RejectsNonReadOnlyActions;
 
     public function view(?User $user): bool
     {
